Move tests and builders to relevant folders
