+++ src/domain/contracts/contracts/__tests__/multi-send-decoder.helper.spec.ts
@@ -7,11 +7,11 @@ import {
   changeThresholdEncoder,
   removeOwnerEncoder,
   swapOwnerEncoder,
-} from '@/domain/contracts/contracts/safe-encoder.builder';
+} from '@/domain/contracts/contracts/__tests__/safe-encoder.builder';
 import {
   multiSendEncoder,
   multiSendTransactionsEncoder,
-} from '@/domain/contracts/contracts/multi-send-encoder.builder';
+} from '@/domain/contracts/contracts/__tests__/multi-send-encoder.builder';
 import { safeBuilder } from '@/domain/safe/entities/__tests__/safe.builder';
 
 describe('MultiSendDecoder', () => {
