+++ src/domain/contracts/contracts/__tests__/safe-encoder.builder.ts
